+++ server/v4a_sender.go
@@ -25,12 +25,12 @@ func NewBaseV4aSender(protocol v4a.Protocol, tcpConfig config.TcpConfig) (BaseV4
 }
 
 func (b BaseV4aSender) SendFailAndClose(client net.Conn) {
-	_ = b.protocol.ResponseWithFail(uint16(b.tcpConfig.GetBindPort()), net.IP{0, 0, 0, 0}, client)
+	_ = b.protocol.ResponseWithFail(b.tcpConfig.GetBindPort(), net.IP{0, 0, 0, 0}, client)
 	_ = client.Close()
 }
 
 func (b BaseV4aSender) SendSuccess(client net.Conn) error {
-	return b.protocol.ResponseWithSuccess(uint16(b.tcpConfig.GetBindPort()), net.IP{0, 0, 0, 0}, client)
+	return b.protocol.ResponseWithSuccess(b.tcpConfig.GetBindPort(), net.IP{0, 0, 0, 0}, client)
 }
 
 func (b BaseV4aSender) SendSuccessWithParameters(ip net.IP, port uint16, client net.Conn) error {
